Ruleset/requirements.txt
polars
orjson
httpx
pyyaml
regex
